Refactor and Rename Extractor (#104)
* Refactor and Rename MultiExtractor to RecursiveExtractor
* Fix Broken Test Data
