+++ src/RecursiveExtractor.Tests/RecursiveExtractor.Tests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -15,7 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MultiExtractor\MultiExtractor.csproj" />
+    <ProjectReference Include="..\RecursiveExtractor\RecursiveExtractor.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Update="TestData\10GB.7z.bz2">
