+++ src/RecursiveExtractor.Tests/TestData/10GB.gz.bz2
