+++ src/RecursiveExtractor.Tests/TestData/Empty.vmdk
