+++ src/RecursiveExtractor.Tests/TestData/Nested.zip
