+++ src/RecursiveExtractor.Tests/TestData/Shared.vhd
