+++ src/RecursiveExtractor.Tests/TestData/Shared.wim
