+++ src/RecursiveExtractor.Tests/TestData/Shared.zip
