+++ src/RecursiveExtractor.Tests/TestData/droste.zip
