+++ src/RecursiveExtractor.Tests/TestData/zbsm.zip
