+++ src/RecursiveExtractor/DebArchiveFile.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Text;
 
-namespace Microsoft.CST.OpenSource.MultiExtractor
+namespace Microsoft.CST.OpenSource.RecursiveExtractor
 {
     /**
      * Very simple implementation of an .Deb format parser, needed for Debian .deb archives.
