+++ src/RecursiveExtractor/MiniMagic.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Text;
 
-namespace Microsoft.CST.OpenSource.MultiExtractor
+namespace Microsoft.CST.OpenSource.RecursiveExtractor
 {
     /// <summary>
     /// ArchiveTypes are the kinds of archive files that this module can process.
